Exercise Removed Defaults + minor changes

+++ lib/settings/screen_settings.dart
@@ -35,22 +35,36 @@ class _SettingScreenState extends ConsumerState<SettingScreen> {
         children: [
           ListTile(
             title: const Text("Delete all Exercise"),
-            onTap: () => ref
-                .watch(exerciseManagerProvider.future)
-                .then((value) => value.deleteAllExercise()),
+            onTap: () async {
+              bool confirm = await showConfirmDialog(context) ?? false;
+              if (confirm) {
+                ref
+                    .watch(exerciseManagerProvider.future)
+                    .then((value) => value.deleteAllExercise());
+              }
+            },
           ),
           ListTile(
             title: const Text("Delete all Workouts"),
-            onTap: () => ref
-                .watch(workoutManagerProvider.future)
-                .then((value) => value.deleteAllWorkout()),
+            onTap: () async {
+              bool confirm = await showConfirmDialog(context) ?? false;
+              if (confirm) {
+                ref
+                    .watch(workoutManagerProvider.future)
+                    .then((value) => value.deleteAllWorkout());
+              }
+            },
           ),
           ListTile(
-            title: const Text("Clear all Data"),
-            onTap: () => ref
-                .watch(isarInstanceProvider.future)
-                .then((value) => value.writeTxn(() => value.clear())),
-          ),
+              title: const Text("Clear all Data"),
+              onTap: () async {
+                bool confirm = await showConfirmDialog(context) ?? false;
+                if (confirm) {
+                  ref
+                      .watch(isarInstanceProvider.future)
+                      .then((value) => value.writeTxn(() => value.clear()));
+                }
+              }),
           SwitchListTile(
             value: _isDarkTheme,
             title: const Text("Dark Theme"),
@@ -67,6 +81,33 @@ class _SettingScreenState extends ConsumerState<SettingScreen> {
           ),
         ],
       ),
+    );
+  }
+
+  Future<bool?> showConfirmDialog(BuildContext context) {
+    return showDialog<bool>(
+      context: context,
+      builder: (BuildContext context) {
+        return AlertDialog(
+          title: const Text("This Cannot Be Undone!\nAre you Sure?"),
+          actions: [
+            TextButton(
+              child: Text(
+                "No",
+                style: Theme.of(context).textTheme.titleSmall,
+              ),
+              onPressed: () => Navigator.of(context).pop(false),
+            ),
+            TextButton(
+              child: Text(
+                "Yes",
+                style: Theme.of(context).textTheme.titleSmall,
+              ),
+              onPressed: () => Navigator.of(context).pop(true),
+            ),
+          ],
+        );
+      },
     );
   }
 }
